requirements-dev.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.25.0
colorama==0.4.4
decorator==5.0.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.3
django-bootstrap-form==3.4
django-cloudinary-storage==0.3.0
django-extensions==3.1.3
django-session-security==2.6.6
django-session-timeout==0.1.0
django-widget-tweaks==1.4.8
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
Pygments==2.9.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
